Optimized RenderTileMap

+++ src/RenderTileMap.h
@@ -6,7 +6,7 @@ class RenderTileMap : public RenderComponent {
     public:
         Texture2D* texture;
         Rectangle tiles[64];
-        int map[TILEMAP_WIDTH][TILEMAP_HEIGHT];
+        short map[TILEMAP_WIDTH][TILEMAP_HEIGHT];
         AABB* bounds;
 
         bool isOnScreen(const AABB*) const override;
